refactor: update the double pane component
